Refactor auth interceptor
